fix(cli-sub-agent): fail closed on uncertain worktree-lock holders (#1873 #1838)
Split registry-absent holder state from loaded terminal holder state so missing or unreadable session state no longer auto-classifies as Dead.

Registry-absent holders are reclaimable only when the diagnostic PID is independently proven missing, preserving #1838 crash-stale recovery while keeping #1672/#1828 write serialization fail-closed for live, unreadable, or otherwise uncertain holders.

Targeted tests cover registry-absent signalable PID blocking, unreadable state blocking, registry-absent ESRCH reclaim, loaded-terminal reclaim, and the existing pipeline write-lock invariant.
